Move mock config to app/config
