+++ app/point/config/testing/mocks/config.go
